requirements.txt
actionlib==1.14.0
angles==1.9.13
base_local_planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.8.0
docutils==0.19
dynamic-reconfigure==1.7.3
dynamixel_sdk==3.7.51
fkie_master_discovery==1.3.2
fkie_master_sync==1.3.2
fkie_node_manager_daemon==1.3.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
haru_unity==0.3.20
image-geometry==1.16.2
interactive-markers==1.12.0
joint-state-publisher==1.15.1
laser_geometry==1.6.7
message-filters==1.16.0
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-qt-binding==0.4.4
python-rtmidi==1.4.9
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
ros_tcp_endpoint==0.6.0
rosapi==0.11.16
rosbag==1.16.0
rosboost-cfg==1.15.8
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospy==1.16.0
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-reconfigure==0.5.5
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_plot==0.4.13
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.19
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
sound-play==0.3.16
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
topic-tools==1.16.0
turtlebot3_autorace_camera==1.1.1
turtlebot3_autorace_core==1.1.1
turtlebot3_autorace_detect==1.1.1
turtlebot3_autorace_driving==1.1.1
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
xacro==1.14.14
